Create useMultisigsList hook. Update multisigs list query logic.
