+++ pages/multisig/list.tsx
@@ -1,10 +1,9 @@
 import React, { FunctionComponent, useEffect, useState } from 'react'
-import { Contract } from '@cosmjs/cosmwasm-stargate'
 import { ChevronRightIcon } from '@heroicons/react/solid'
 import WalletLoader from 'components/WalletLoader'
-import { useSigningClient } from 'contexts/cosmwasm'
 import type { NextPage } from 'next'
 import Link from 'next/link'
+import { useMultisigsList } from 'hooks/multisig'
 import { MULTISIG_CODE_ID } from 'util/constants'
 
 interface MultisigListType {
@@ -31,39 +30,7 @@ const MultisigListComponent: FunctionComponent<MultisigListType> = ({
 }
 
 const MultisigList: NextPage = () => {
-  const [multisigs, setMultisigs] = useState<Array<Contract>>([])
-  const { signingClient } = useSigningClient()
-  const [loading, setLoading] = useState(false)
-
-  // Get list of MULTISIG info
-  useEffect(() => {
-    const getMultisigs = async () => {
-      setLoading(true)
-      try {
-        let contracts = await signingClient?.getContracts(MULTISIG_CODE_ID)
-
-        let multisigList: Array<Contract> = []
-        if (contracts) {
-          for (let address of contracts) {
-            signingClient?.getContract(address).then((response) => {
-              multisigList.push(response)
-            })
-          }
-          if (multisigList.length > 0) {
-            setMultisigs(multisigList)
-          }
-          setLoading(false)
-        }
-      } catch (e) {
-        // Handles the edge case where there are no contracts at
-        // all and the API throws.
-        setLoading(false)
-        return []
-      }
-    }
-    getMultisigs()
-  }, [signingClient])
-
+  const { multisigs, loading } = useMultisigsList(MULTISIG_CODE_ID)
   return (
     <WalletLoader loading={loading}>
       <h1 className="text-6xl font-bold">MULTISIGs</h1>
